refactor(control-server): split monolithic entrypoint into focused modules
`src/index.ts` had grown to ~1030 lines mixing Fastify bootstrap, auth,
two WebSocket protocols, inline invite-exchange HTML/JS, Yjs broadcast, and
startup logging. Protocol changes meant editing the monolith and the inline
invite script could not be linted or unit-tested.

Extract the pieces into cohesive modules, threading shared connection state
and dependencies through an `AppContext` passed by reference:

- config.ts        — env parsing, constants, resolveListenPort, parseTrustProxy
- context.ts       — Client/StreamwallConnection/AppContext + broadcast factory
- wsSupport.ts     — bearerToken, queueWebSocketMessages, message-rate guard
- inviteExchange/  — page.html + exchange.js (served from disk) + invite routes
- ws/uplink.ts     — desktop Streamwall uplink handler
- ws/client.ts     — authenticated plugin (auth, /admin/status, static, client ws)
- bootstrap.ts     — initialInviteCodes, logBootstrap, runServer

index.ts is now ~200 lines of app assembly plus the process entry guard, and
re-exports the public surface so existing importers and the test suite keep a
single stable entry point. No protocol behavior changes.

The invite-exchange script moves from an inline string to a real ES module
served with `<script type="module">` (still same-origin, so it satisfies the
strict `script-src 'self'` CSP), read from disk alongside its page.html.

Refs #395
