+++ packages/streamwall-control-server/src/inviteExchange/page.html
@@ -1,0 +1,12 @@
+<!doctype html>
+<html lang="en">
+  <head>
+    <meta charset="utf-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1" />
+    <title>Streamwall — joining…</title>
+  </head>
+  <body>
+    <p>Signing you in…</p>
+    <script type="module" src="/invite-exchange.js"></script>
+  </body>
+</html>
